backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
pydantic
sqlalchemy
psycopg2-binary
alembic
python-multipart